% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosign.R
\docType{methods}
\name{biosign,MultiDataSet-method}
\alias{biosign,MultiDataSet-method}
\title{Feature selection}
\usage{
\S4method{biosign}{MultiDataSet}(x, y, seedI = NULL, fig.pdfC = NA,
  plotTierMaxC = "S", info.txtC = NA, ...)
}
\arguments{
\item{x}{An S4 object of class \code{MultiDataSet}}

\item{y}{a character indicating the name of the column of the pData
to be used, when x is an ExpressionSet object}

\item{seedI}{integer: optional seed to obtain exactly the same signature when rerunning biosigner}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen; if NULL, no plot is generated}

\item{plotTierMaxC}{Character: Maximum level of tiers to display: Either 'S' and
'A', (for boxplot), or also 'B', 'C', 'D', and 'E' (for tiers) by decreasing
number of selections}

\item{info.txtC}{Character: File name for the printed results (call to 'sink()')}

\item{...}{additional parameters for the ropls::opls modeling}
}
\value{
list of 'mset', the \code{MultiDataSet} including the tiers in fData 
and 'signLs', the list of 'biosign' outputs
}
\description{
Feature selection
}
\examples{
prometMset <- readSet(system.file("extdata/promet", package="metabolis"))
## selecting a seed (for reproducibility) and a low bootstrap number (for speed in this demo)
\dontrun{
set.seed(123)
prometMset <- biosign(prometMset, "gene", bootI = 5)
## viewing selected features:
lapply(fData(prometMset), function(fda) {
  tierDF <- fda[, grep("biosign", colnames(fda))]
  tierDF[rowSums(tierDF == "S") > 0, ]
})
}
}
