% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolis-package.R
\docType{package}
\name{metabolis-package}
\alias{metabolis-package}
\alias{metabolis}
\title{Metabolomics Data Sciences and Integration}
\description{
Package description
}
\examples{
prometMset <- metRead(system.file("extdata/promet", package="metabolis"))
\dontrun{
prometMset <- metView(prometMset)
prometMset <- metView(prometMset, factorC = "gene")
}

}
\author{
E. A. Thévenot (CEA, LIST, MetaboHUB)

Maintainer: Etienne Thévenot <etienne.thevenot@cea.fr>
}
\keyword{package}
