% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metCorrect.R
\docType{methods}
\name{metCorrect}
\alias{metCorrect}
\alias{metCorrect,MultiDataSet-method}
\alias{metCorrect,ExpressionSet-method}
\title{Signal drift and batch effect correction}
\usage{
metCorrect(x, ...)

\S4method{metCorrect}{MultiDataSet}(x, referenceSampleTypeC = c("pool",
  "sample")[1], colnameBatchC = "batch",
  colnameInjectionOrderC = "injectionOrder",
  colnameSampleTypeC = "sampleType", spanN = 1, fig.pdfC = NA,
  info.txtC = NA)

\S4method{metCorrect}{ExpressionSet}(x, referenceSampleTypeC = c("pool",
  "sample")[1], colnameBatchC = "batch",
  colnameInjectionOrderC = "injectionOrder",
  colnameSampleTypeC = "sampleType", spanN = 1, fig.pdfC = NA,
  info.txtC = NA)
}
\arguments{
\item{x}{An S4 object of class \code{MultiDataSet}}

\item{colnameBatchC}{Character: name of the column from pData(x) containing the batch information (encoded as characters)}

\item{colnameSampleTypeC}{Character:  name of the column from pData(x) containing the sample type information (encoded as characters)}

\item{spanN}{Numeric: smoothing parameter for the loess regression; between 0 and 1; (default set to 1)}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen;}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}

\item{referenceSampeTypeC}{Character: sample type to be used as reference for the correction (as indicated in the 'colnameSampleType' column from the
pData(x); e.g. 'pool')}

\item{colnameInjectionOrder}{Character: name of the column from pData(x) containing the injection order information (encoded as numerics)}

\item{x}{An S4 object of class \code{ExpressionSet}}

\item{referenceSampeTypeC}{Character: sample type to be used as reference for the correction (as indicated in the 'colnameSampleType' column from the
pData(x); e.g. 'pool')}

\item{colnameBatchC}{Character: name of the column from pData(x) containing the batch information (encoded as characters)}

\item{colnameInjectionOrder}{Character: name of the column from pData(x) containing the injection order information (encoded as numerics)}

\item{colnameSampleTypeC}{Character:  name of the column from pData(x) containing the sample type information (encoded as characters)}

\item{spanN}{Numeric: smoothing parameter for the loess regression; between 0 and 1; (default set to 1)}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen;}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
\code{MultiDataSet} including the adjusted p-values in fData (or the vector of adjusted p-values when returnAdjustOnlyL is TRUE)

\code{ExpressionSet} including the adjusted p-values in fData (or the vector of adjusted p-values when returnAdjustOnlyL is TRUE)
}
\description{
Signal drift and batch effect correction

Signal drift and batch effect correction
}
\examples{
# to be provided
sacSet <- metRead(system.file("extdata/sacurine", package = "metabolis"))
sacSet <- metCorrect(sacSet)
}
