% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metHeatmap.R
\docType{methods}
\name{metHeatmap}
\alias{metHeatmap}
\alias{metHeatmap,MultiDataSet-method}
\alias{metHeatmap,ExpressionSet-method}
\title{Heatmap}
\usage{
metHeatmap(x, ...)

\S4method{metHeatmap}{MultiDataSet}(x, fig.pdfC = NA, info.txtC = NA,
  ...)

\S4method{metHeatmap}{ExpressionSet}(x, dissymC = c("euclidean",
  "maximum", "manhattan", "canberra", "binary", "minkowski", "1-cor",
  "1-abs(cor)")[7], correlC = c("pearson", "kendall", "spearman"),
  aggloC = c("ward.D", "ward.D2", "single", "complete", "average",
  "mcquitty", "median", "centroid")[2], clustVi = c(2, 2),
  fig.pdfC = NA, plotCexVn = c(1, 1),
  plotPaletteC = c("blueOrangeRed", "redBlackGreen")[1],
  plotScaleL = TRUE, info.txtC = NA)
}
\arguments{
\item{x}{an S4 object of class \code{MultiDataSet}}

\item{...}{additional parameters (see the documentation)}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen;}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}

\item{dissymC}{Character: [hclust]}

\item{correlC}{Character: correlation coefficient (in case
'1-cor' or '1-abs(cor)' are selected as dissymilarity)}

\item{aggloC}{charcter: agglomeration method}

\item{clustVi}{tupple of integers: number of sample and variable clusters, respectively}

\item{plotCexVn}{vector of numerics:}

\item{plotPaletteC}{character: color palette}

\item{plotScaleL}{logical: scaling (mean-centering and unit
variance scaling) to enhance contrast (for plotting only)}

\item{x}{An S4 object of class \code{ExpressionSet}}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen;}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
\code{MultiDataSet} including the column with the
cluster numbers in pData and fData

\code{ExpressionSet} including the adjusted p-values in fData (or the vector of adjusted p-values when returnAdjustOnlyL is TRUE)
}
\description{
Heatmap of a MultiDataSet instance

Heatmap of an ExpressionSet instance
}
\examples{
prometMset <- readSet(system.file("extdata/promet/",
package="metabolis"))
prometMset <- metHeatmap(prometMset)
metSet <- readSet(system.file("extdata/promet/metabo",
package="metabolis"))
metSet <- metHeatmap(metSet)
head(fData(metSet))
}
