% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metRead-Write.R
\name{metRead}
\alias{metRead}
\title{metRead}
\usage{
metRead(dirC, filesLs = NULL, subsetVc = NA, info.txtC = NA)
}
\arguments{
\item{dirC}{Character: directory containing the 3 .tsv files (single
dataset), or containing several subdirectories with 3 .tsv files
(multiple datasets)}

\item{filesLs}{List: if dirC is set to NA, the full names of the
individual files can be provided; in case of an ExpressionSet, the
names of the list must be 'dataMatrix.tsvC', 'sampleMetadata.tsvC',
and 'variableMetadata.tsvC' with the corresponding file full names;
in case of a MultiDataSet, the list must consists of one such sublist
per dataset}

\item{subsetVc}{Vector of characters: specifying a subset of the
subdirectories to be included in the MultiDataSet (by default, all
subdirectories containing the 3 tables will be considered as datasets)}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
MultiDataSet (multiple dataset) instance
}
\description{
Reading datasets as subfolders containing the 3 tables 'dataMatrix.tsv',
'sampleMetadata.tsv' and 'variableMetadata.tsv'
}
\examples{
prometDirC <- system.file("extdata/promet", package="metabolis")
## 1) Single set
metEset <- metRead(file.path(prometDirC, "metabo"))
# or
metEset <- metRead(NA,
                    filesLs = list(dataMatrix.tsvC = file.path(prometDirC, "metabo", "dataMatrix.tsv"),
                                   sampleMetadata.tsvC = file.path(prometDirC, "metabo", "sampleMetadata.tsv"),
                                   variableMetadata.tsvC = file.path(prometDirC, "metabo", "variableMetadata.tsv")))
## 2) Multiple sets
prometMset <- metRead(prometDirC)
metMset <- metRead(prometDirC, subsetVc = "metabo")
# or
prometMset <- metRead(NA,
                       filesLs = list(metabo = list(dataMatrix.tsvC = file.path(prometDirC, "metabo", "dataMatrix.tsv"),
                                                    sampleMetadata.tsvC = file.path(prometDirC, "metabo", "sampleMetadata.tsv"),
                                                    variableMetadata.tsvC = file.path(prometDirC, "metabo", "variableMetadata.tsv")),
                                      proteo = list(dataMatrix.tsvC = file.path(prometDirC, "proteo", "dataMatrix.tsv"),
                                                    sampleMetadata.tsvC = file.path(prometDirC, "proteo", "sampleMetadata.tsv"),
                                                    variableMetadata.tsvC = file.path(prometDirC, "proteo", "variableMetadata.tsv"))))
}
\author{
Etienne Thevenot, \email{etienne.thevenot@cea.fr}
}
