% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metTest.R
\docType{methods}
\name{metTest}
\alias{metTest}
\alias{metTest,MultiDataSet-method}
\alias{metTest,ExpressionSet-method}
\title{Univariate hypothesis testing}
\usage{
metTest(x, ...)

\S4method{metTest}{MultiDataSet}(x, factorC, testC = c("param",
  "nonparam", "ttest", "limma", "wilcoxon", "anova", "kruskal", "pearson",
  "spearman")[2], adjustC = c("holm", "hochberg", "hommel", "bonferroni",
  "BH", "BY", "fdr", "none")[5], adjustThreshN = 0.05, fig.pdfC = NA,
  info.txtC = NA)

\S4method{metTest}{ExpressionSet}(x, factorC, testC = c("param",
  "nonparam", "ttest", "limma", "wilcoxon", "anova", "kruskal", "pearson",
  "spearman")[2], adjustC = c("holm", "hochberg", "hommel", "bonferroni",
  "BH", "BY", "fdr", "none")[5], adjustThreshN = 0.05, fig.pdfC = NA,
  plotMainC = NA, info.txtC = NA)
}
\arguments{
\item{x}{An S4 object of class \code{MultiDataSet}}

\item{factorC}{Character: Factor of interest (name of a column from the
pData(x))}

\item{testC}{Character: One of the 6 available hypothesis tests can be selected
(either 'ttest', 'limma', 'wilcoxon', 'anova', 'kruskal', 'pearson', or 'spearman');
if set to 'param' (or 'nonparam' [default]), the appropriate (non-)parametric test will be used}

\item{adjustC}{Character: Name of the method for correction of multiple testing
(the p.adjust function is used)}

\item{adjustThreshN}{Numeric: Threshold for (corrected) p-values}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen; if NULL, no plot is generated}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}

\item{plotMainC}{character: title to be displayed on the first plot}

\item{x}{An S4 object of class \code{ExpressionSet}}

\item{factorC}{Character: Factor of interest (name of a column from the
pData(x))}

\item{testC}{Character: One of the 6 available hypothesis tests can be selected
(either 'ttest', 'limma', 'wilcoxon', 'anova', 'kruskal', 'pearson', or 'spearman');
if set to 'param' (or 'nonparam' [default]), the appropriate (non-)parametric test will be used}

\item{adjustC}{Character: Name of the method for correction of multiple testing
(the p.adjust function is used)}

\item{adjustThreshN}{Numeric: Threshold for (corrected) p-values}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen; if NULL, no plot is generated}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
\code{MultiDataSet} including the adjusted p-values in the fData data frames

\code{ExpressionSet} including the adjusted p-values in fData (or the vector of adjusted p-values when returnAdjustOnlyL is TRUE)
}
\description{
Provides numerical metrics and graphical overview of an ExpressionSet instance

Provides numerical metrics and graphical overview of an ExpressionSet instance
}
\examples{
prometMset <- readSet(system.file("extdata/promet", package="metabolis"))
prometMset <- metTest(prometMset, "gene")
proSet <- readSet(system.file("extdata/promet/proteo", package="metabolis"))
proSet <- metTest(proSet, "gene", fig.pdfC = NULL)
head(fData(proSet))
\dontrun{
proSet <- metTest(proSet, "gene", fig.pdfC = NA)
proSet <- metTest(proSet, "gene", fig.pdfC = "proSet_metTest-gene.pdf")
}
}
