% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metTransform.R
\docType{methods}
\name{metTransform}
\alias{metTransform}
\alias{metTransform,MultiDataSet-method}
\alias{metTransform,ExpressionSet-method}
\title{Transformation of the dataMatrix}
\usage{
metTransform(x, ...)

\S4method{metTransform}{MultiDataSet}(x, methodC = c("log2", "log10",
  "sqrt")[1], info.txtC = NA)

\S4method{metTransform}{ExpressionSet}(x, methodC = c("log2", "log10",
  "sqrt")[1], info.txtC = NA)
}
\arguments{
\item{x}{An S4 object of class \code{MultiDataSet}}

\item{methodC}{Character: Factor of interest (name of a column from the
pData(x))}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}

\item{x}{An S4 object of class \code{ExpressionSet}}

\item{methodC}{Character: Factor of interest (name of a column from the
pData(x))}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
\code{MultiDataSet} including the adjusted p-values in the fData data frames

\code{ExpressionSet} including the adjusted p-values in fData (or the vector of adjusted p-values when returnAdjustOnlyL is TRUE)
}
\description{
Transformation of the dataMatrix

Transformation of the dataMatrix
}
\examples{
prometMset <- metRead(system.file("extdata/promet", package="metabolis"))
prometMset <- metTransform(prometMset, "sqrt")
printStr <- lapply(assayData(prometMset), function(aData) ropls::strF(aData[["exprs"]]))
sacSet <- metRead(system.file("extdata/sacurine", package = "metabolis"))
ropls::strF(t(exprs(sacSet)))
sacSet <- metTransform(sacSet, "log2")
ropls::strF(t(exprs(sacSet)))
}
