% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/metView.R
\docType{methods}
\name{metView}
\alias{metView}
\alias{metView,MultiDataSet-method}
\alias{metView,ExpressionSet-method}
\title{Viewing the MultiDataSet object.}
\usage{
metView(x, ...)

\S4method{metView}{MultiDataSet}(x, factorC = NA, fig.pdfC = NA,
  poolAsPool1L = FALSE, poolCvN = 0.3, univTestC = "nonparam",
  univAdjustC = "BH", univThreshN = 0.05, plotMainC = "",
  plotOverviewL = TRUE, plotSpanN = 1,
  plotSampIntensityC = c("median", "mean")[1], info.txtC = NA)

\S4method{metView}{ExpressionSet}(x, factorC = NA, fig.pdfC = NA,
  poolAsPool1L = FALSE, poolCvN = 0.3, univTestC = "nonparam",
  univAdjustC = "BH", univThreshN = 0.05, plotMainC = NA,
  plotSpanN = 1, plotSampIntensityC = c("median", "mean")[1],
  info.txtC = NA)
}
\arguments{
\item{x}{MultiDataSet: multiple dataset to be explored}

\item{factorC}{Character: column name of pData to be used for sample color and
univariate test}

\item{fig.pdfC}{Character: Name of the file for the graphics; if NA, the
graphics are displayed on the screen;}

\item{poolAsPool1L}{Logical: should pool be included (as pool1) in the correlation
with the dilution factor?}

\item{poolCvN}{Numeric: threshold for the coefficient of variation of the pools}

\item{univTestC}{Character: if 'factorC' is specified, which test should be used
(see the univariate method)}

\item{univAdjustC}{Character: if 'factorC' is specified, which correction for
multiple testing should be used (see the univariate method)}

\item{plotMainC}{Character: title; note that all 'plot' parameters are used only
for display}

\item{plotOverviewL}{Logical: should an overview of the number of samples and
variables in all datasets be barplotted?}

\item{plotSpanN}{Numeric: span parameter used in the loess trend estimation}

\item{plotSampIntensityC}{Character: function to be used to display the global
sample intensity}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}

\item{univTreshN}{Character: if 'factorC' is specified, which threshold for
the corrected p-values should be used to reject the null hypothesis?}

\item{x}{ExpressionSet: dataset to be explored}

\item{factorC}{Character: column name of pData to be used for sample color and
univariate test}

\item{fig.pdfC}{Character: name of the pdf file for output}

\item{poolAsPool1L}{Logical: should pool be included (as pool1) in the correlation
with the dilution factor?}

\item{poolCvN}{Numeric: threshold for the coefficient of variation of the pools}

\item{univTestC}{Character: if 'factorC' is specified, which test should be used
(see the univariate method)}

\item{univAdjustC}{Character: if 'factorC' is specified, which correction for
multiple testing should be used (see the univariate method)}

\item{univTreshN}{Character: if 'factorC' is specified, which threshold for
the corrected p-values should be used to reject the null hypothesis?}

\item{plotMainC}{Character: title; note that all 'plot' parameters are used only for display}

\item{plotSpanN}{Numeric: span parameter used in the loess trend estimation}

\item{plotSampIntensityC}{Character: function to be used to display the global
sample intensity}

\item{info.txtC}{Character: File name for the printed results (call to
'sink()'); if NA (default), messages will be printed on the screen; if NULL,
no verbose will be generated}
}
\value{
MultiDataSet including the computed sample and variable metrics

ExpressionSet including the computed sample and variable metrics
}
\description{
Provides numerical metrics and graphical overview of each dataset

Provides numerical metrics and graphical overview of an ExpressionSet instance
}
\examples{
prometMset <- metRead(system.file("extdata/promet",
                                  package = "metabolis"))
\dontrun{
prometMset <- metView(prometMset)
prometMset <- metView(prometMset, factorC = "gene")
}

metSet <- metRead(system.file("extdata/promet/metabo",
                              package = "metabolis"))
\dontrun{
metSet <- metView(metSet)
metSet <- metView(metSet, factorC = "gene")
}

}
