% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls.R
\docType{methods}
\name{opls,MultiDataSet-method}
\alias{opls,MultiDataSet-method}
\title{PCA and (O)PLS-DA modeling}
\usage{
\S4method{opls}{MultiDataSet}(x, y = NULL, fig.pdfC = NA,
  info.txtC = NULL, ...)
}
\arguments{
\item{x}{An S4 object of class \code{MultiDataSet}}

\item{y}{Character: Name of the fData column to be used for (O)PLS(-DA) modeling;
(set to NULL in case of PCA)}

\item{fig.pdfC}{Character: Name of the file for the graphics from the significant features;
if NA, the graphics are displayed on the screen; if NULL, no plot is generated}

\item{info.txtC}{Character: File name for the printed results (call to 'sink()')}

\item{...}{additional parameters for the ropls::opls modeling}
}
\value{
list of 'mset', the \code{MultiDataSet} including the scores, loadings,
VIP, etc. in the corresponding pData and fData when a model could be built
(>=1 component) and was significant (pQ2 <= 5%), and 'oplsLs', the list
of 'opls' outputs
}
\description{
PCA and (O)PLS-DA modeling
}
\examples{
prometMset <- readSet(system.file("extdata/promet", package="metabolis"))
prometMset <- opls(prometMset, "gene", predI = 1, orthoI = 1)
}
