% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metView.R
\name{vennF}
\alias{vennF}
\title{Venn Diagram}
\usage{
vennF(inputLs, file.tiffC = NULL, lwdN = 2, mainC = NA,
  palVc = RColorBrewer::brewer.pal(9, "Set1")[1:5], subC = "")
}
\arguments{
\item{inputLs}{list containing up to 5 vectors}

\item{file.tiffC}{file name for the figure, if set to
NULL (default), the figure is displayed on the screen}

\item{lwdN}{size of the circle lines}

\item{mainC}{title (by default, the name of the matrixMN variable will be used)}

\item{palVc}{color palette to be used}

\item{subVc}{subtitle}
}
\value{
No output.
}
\description{
Venn Diagram of a list containing up to 5 vectors
}
\examples{
print("hello")
}
